+++ helpers.h
@@ -7,7 +7,7 @@ using namespace std;
 
 const unordered_set<string> INS = {"add"}; //TODO: finish if makes sense
 
-void outWord(const int64_t& word);
+void outWord(const int& word);
 
 //! REMOVE
 unordered_map<string, int> getSymbolTable(const vector<Token>& v);
